requirements.txt
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.115.0
filelock==3.16.1
fsspec==2024.9.0
grpcio==1.66.2
grpcio-tools==1.66.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.1
hyperframe==6.0.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
joblib==1.4.2
loguru==0.7.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
nltk==3.9.1
numpy==2.1.1
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
portalocker==2.10.1
protobuf==5.28.2
pydantic==2.9.2
pydantic_core==2.23.4
PyPDF2==3.0.1
pytest==8.3.3
python-dotenv==1.0.1
PyYAML==6.0.2
qdrant-client==1.11.3
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.1.1
sniffio==1.3.1
starlette==0.38.6
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.0
torch==2.4.1
tqdm==4.66.5
transformers==4.45.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
